run scalafmt on modifed block only and move to sql package to handle review comment
