+++ sql/core/src/main/scala/org/apache/spark/sql/execution/ui/SQLAppStatusStore.scala
@@ -34,7 +34,9 @@ import org.apache.spark.util.kvstore.{KVIndex, KVStore}
  * Provides a view of a KVStore with methods that make it easy to query SQL-specific state. There's
  * no state kept in this class, so it's ok to have multiple instances of it in an application.
  */
-class SQLAppStatusStore(store: KVStore, val listener: Option[SQLAppStatusListener] = None) {
+class SQLAppStatusStore(
+                         store: KVStore,
+                         val listener: Option[SQLAppStatusListener] = None) {
 
   def executionsList(): Seq[SQLExecutionUIData] = {
     store.view(classOf[SQLExecutionUIData]).asScala.toSeq
@@ -80,33 +82,33 @@ class SQLAppStatusStore(store: KVStore, val listener: Option[SQLAppStatusListene
 }
 
 class SQLExecutionUIData(
-    @KVIndexParam val executionId: Long,
-    val description: String,
-    val details: String,
-    val physicalPlanDescription: String,
-    val metrics: Seq[SQLPlanMetric],
-    val submissionTime: Long,
-    val completionTime: Option[Date],
-    @JsonDeserialize(keyAs = classOf[Integer])
-    val jobs: Map[Int, JobExecutionStatus],
-    @JsonDeserialize(contentAs = classOf[Integer])
-    val stages: Set[Int],
-    /**
-     * This field is only populated after the execution is finished; it will be null while the
-     * execution is still running. During execution, aggregate metrics need to be retrieved
-     * from the SQL listener instance.
-     */
-    @JsonDeserialize(keyAs = classOf[JLong])
-    val metricValues: Map[Long, String]) {
+                          @KVIndexParam val executionId: Long,
+                          val description: String,
+                          val details: String,
+                          val physicalPlanDescription: String,
+                          val metrics: Seq[SQLPlanMetric],
+                          val submissionTime: Long,
+                          val completionTime: Option[Date],
+                          @JsonDeserialize(keyAs = classOf[Integer])
+                          val jobs: Map[Int, JobExecutionStatus],
+                          @JsonDeserialize(contentAs = classOf[Integer])
+                          val stages: Set[Int],
+                          /**
+                           * This field is only populated after the execution is finished; it will be null while the
+                           * execution is still running. During execution, aggregate metrics need to be retrieved
+                           * from the SQL listener instance.
+                           */
+                          @JsonDeserialize(keyAs = classOf[JLong])
+                          val metricValues: Map[Long, String]) {
 
   @JsonIgnore @KVIndex("completionTime")
   private def completionTimeIndex: Long = completionTime.map(_.getTime).getOrElse(-1L)
 }
 
 class SparkPlanGraphWrapper(
-    @KVIndexParam val executionId: Long,
-    val nodes: Seq[SparkPlanGraphNodeWrapper],
-    val edges: Seq[SparkPlanGraphEdge]) {
+                             @KVIndexParam val executionId: Long,
+                             val nodes: Seq[SparkPlanGraphNodeWrapper],
+                             val edges: Seq[SparkPlanGraphEdge]) {
 
   def toSparkPlanGraph(): SparkPlanGraph = {
     SparkPlanGraph(nodes.map(_.toSparkPlanGraphNode()), edges)
@@ -115,17 +117,14 @@ class SparkPlanGraphWrapper(
 }
 
 class SparkPlanGraphClusterWrapper(
-    val id: Long,
-    val name: String,
-    val desc: String,
-    val nodes: Seq[SparkPlanGraphNodeWrapper],
-    val metrics: Seq[SQLPlanMetric]) {
+                                    val id: Long,
+                                    val name: String,
+                                    val desc: String,
+                                    val nodes: Seq[SparkPlanGraphNodeWrapper],
+                                    val metrics: Seq[SQLPlanMetric]) {
 
   def toSparkPlanGraphCluster(): SparkPlanGraphCluster = {
-    new SparkPlanGraphCluster(
-      id,
-      name,
-      desc,
+    new SparkPlanGraphCluster(id, name, desc,
       new ArrayBuffer() ++ nodes.map(_.toSparkPlanGraphNode()),
       metrics)
   }
@@ -134,8 +133,8 @@ class SparkPlanGraphClusterWrapper(
 
 /** Only one of the values should be set. */
 class SparkPlanGraphNodeWrapper(
-    val node: SparkPlanGraphNode,
-    val cluster: SparkPlanGraphClusterWrapper) {
+                                 val node: SparkPlanGraphNode,
+                                 val cluster: SparkPlanGraphClusterWrapper) {
 
   def toSparkPlanGraphNode(): SparkPlanGraphNode = {
     assert(node == null ^ cluster == null, "Exactly one of node, cluster values to be set.")
@@ -144,4 +143,7 @@ class SparkPlanGraphNodeWrapper(
 
 }
 
-case class SQLPlanMetric(name: String, accumulatorId: Long, metricType: String)
+case class SQLPlanMetric(
+                          name: String,
+                          accumulatorId: Long,
+                          metricType: String)
